Here is an example of code, creation of a template, then modification of the title for example. (The goal would be to modify the metadata but it is more readable with the title).
This shows me an error, whether or not I put the body argument, or other.

By going into the details of the patch_experiment_template() function, I could see that this function only takes id as argument. Does the error come from the ? Or is it me who misunderstood the use of the function?